engine: make Angle a real type instead of a u32 alias
`Angle` was `pub type Angle = u32`, so the compiler could not tell a binary
angle from any other integer and every angle computation was spelled with
`wrapping_add`/`wrapping_sub` plus `as Angle` casts (ANG90 was even an i32
while ANG180/ANG270 were u32).

`Angle(u32)` is now a `#[repr(transparent)]` newtype with wrapping `Add`,
`Sub`, `Neg`, `Mul<u32>`, `Div<u32>`, and three accessors:
  - `fine()`      the FINESINE/FINECOSINE index (`>> ANGLETOFINESHIFT`)
  - `to_signed()` the two's-complement view used for "which side" tests
  - `to_bits()`   the raw value (for tables, shifts and hex output)
ANG45/90/180/270, TRACEANGLE, FATSPREAD and ANG5 are typed constants.
Angle-valued i32 fields (viewangleoffset, rw_angle1) became Angle, and
TANTOANGLE is read through `tan_to_angle()`.

Effect: ~140 `wrapping_*` calls became `+`/`-`, 67 `as Angle` casts and the
`(x >> ANGLETOFINESHIFT) as usize` idiom are gone, and several places that
reused one variable first as an angle and then as a table index now use two
(r_bsp, p_map, p_mobj, p_user, s_sound, p_inter).

Verification: 57 existing tests + 4 new Angle tests pass, goldens untouched;
workspace clippy 0 warnings; timedemo demo1 instruction count identical to
origin/main within 0.006% (16.793G vs 16.792G), cycles within noise.

+++ rust-doomgeneric/engine/src/d_net.rs
@@ -97,7 +97,7 @@ fn init_connect_data(state: &mut GameState, connect_data: &mut NetConnectData) {
         connect_data.drone = true;
     }
     if state.game.options.right {
-        state.render.r_main.viewangleoffset = ANG270 as i32;
+        state.render.r_main.viewangleoffset = ANG270;
         connect_data.drone = true;
     }
     connect_data.gamemode = state.game.doomstat.gamemode as i32;
